An aftermath of the merge from ShizuruNotes. Modify strings for better translation. Show Hatsune with event banners. Print skill equations with value. Show skill equations on Enemy. Calendar filter view (beta)

+++ app/src/main/java/com/github/malitsplus/shizurunotes/ui/charalist/CharaListViewModel.kt
@@ -158,8 +158,8 @@ class CharaListViewModel(
                     valueB = b.charaProperty.getMagicDef()
                 }
                 "8" -> {
-                    valueA = a.charaProperty.getHp()
-                    valueB = b.charaProperty.getHp()
+                    valueA = a.charaProperty.getHp().toInt()
+                    valueB = b.charaProperty.getHp().toInt()
                 }
                 "9" -> {
                     valueA = a.charaProperty.effectivePhysicalHP
